requirements.txt
fastapi==0.135.2
google-genai==1.68.0
notion-client==3.0.0
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
uvicorn==0.42.0
watchdog==6.0.0
